paq/requirement.txt
Django==1.8.5
psycopg2==2.6
Pillow==2.8.1
djangorestframework==3.1
djoser==0.3.1
django-extensions==1.5.5
django-rest-framework-social-oauth2==0.0.8
django-cors-headers==1.1.0

